Renamed Dash Pad and updated the mesh
Dash pads are now boost pads. Changed the mesh to the imported one
instead of a static flattened cube so it was easier to find the correct
orientation of the object in the the scene so I didn't have to cheat the
math to find the forward direction. If the mesh changes in the future
the forward direction should be the same.