Damage, Ring loss, and Death
Added some base logic to SonicBP for when characters take damage. Ring
loss is adjustable with a blueprint pin, Sonic gets knocked back and
input is interrupted for a short time. The level resets when you get hit
with no rings and 'die' but there is no life system just yet.

Additionally made a 'testEnemy' actor that damages SonicBP on overlap. I
also fixed a few redundant things in Dashpad because I can't go a commit
without touching that thing.