Damage Invincibility and 'Blinking' Effect
Added the first iteration of a short invulnerability period after taking
damage. It's set up so that any future invincibility items like the
monitor power up can use the 'isInvulnerable?' variable without
interfering with post hit damage invulnerability. Additionally added a
first pass of the blinking effect seen after Sonic characters are
damaged.  The effect uses a grace period where after the blinking ends
the character is still invincible for a small period of time. This is
also set up so that an invincibility monitor can make use of it in the
future.

Test_UI got modified to change the locations of the elements to better
reflect the positioning seen in an actual Sonic game rather then where
ever I felt like placing them when I originally added the UI.

testEnemy got modified to use the Hit event instead of the overlap
event. This allows it to keep blocking Sonic instead of allowing him to
pass through it while Sonic is invincible. Known issues that need fixed
are that Sonic keeps getting knocked back in the death state if he keeps
touching it and Sonic won't take damage until moving again if
invulnerability ended while still touching the testEnemy mesh. The
latter is probably a side effect from changing from overlap to hit
event, which uses physics to determine if you hit the object or not.