Dropping Rings on damage.
Made a child of Golden Rings to act as the rings dropped when Sonic gets
hit. This first pass includes changes to SonicBP for the logic to spawn
X amount of rings the character is carrying at the moment up to a
maximum. The dropped_rings actors currently spawn with a random X/Y
impulse in a given range and eventually disappear after a time period
set in the object configs. They also have two stages of the blinking cue
to signify how close they're getting to disappearing. You can pick them
up again after a short delay when dropped to make sure the character
doesn't instantly pick them up as they spawn. Currently, the rings don't
remain upright when landing on the ground and the blueprint logic in
both SonicBP and the Dropping_Rings are a mess of uncommented jargon
which needs fixing.