Dropped_Ring polish, SonicBP fixes, commenting
Dropped Rings and Gold Rings have been renamed to be singular instead of
plural. Drop rings now only rotate on one axis and don't simulate the
rest of the physics and flop on the ground like before.  The logic is
now dressed up with variables and a few logic fixes were applied to make
sure that the variables would use one another. Lots of commenting and
clean up on the logic and vars in SonicBP and Dropped_Ring. Some fixes
in SonicBP for an invincibility exploit and continuing to overlap and
collide on death were made. The damage blink effect was tweaked so now
the character only starts blinking after the control delay ends.