Gaining lives with rings, renamed game_level_instance
-Made a 1up event in SonicBP that adds one to the life counter and a
check that calls that event when the character surpasses ring counts of
100. The threshold increases every time you gain a life without dying or
resetting the map, so that if you have already gained a life from
collecting 100 rings, you will need 200 to gain another. Losing rings
and getting it back to the previous threshold doesn't add any lives to
the counter. Additionally the check calls itself again after increasing
the threshold, so if for some reason the character were to gain enough
rings to cross multiple thresholds in a single instant like, 200 or 400
rings at one time, it would iterate and add the appropriate amount of
lives.

-At the moment the check is an event in and of itself. I changed
Gold_Ring in order to call it after every ring pickup.

-DEBUG_Super_Ring is a child of Gold_Ring that I used to check the
functionality of lives. They just give rings in intervals of 100 instead
of 1. Scaled them to be different. As soon as the lives functionality is
tested I'm getting rid of it.

-Renamed Game_Level_Instance to Persistent_Instance_Data because the
former sounded stupid and I don't know what I was thinking. It's easier
to understand what it is meant for named this way. It's just a royal
pain to replace every little reference to it when you change the name,
so I did it right this time. Hopefully....

-Test UI was changed to show a higher amount of rings without being cut
off

-HDKDefaultMap had the debug super rings in it now so that changed

-Ini change to add persistent Instance Data as the game instance. BE
WARNED that changing or renaming game instances names without resets the
game mode to nothing. If you try to change the instance afterward from
the drop down window without resetting the game mode it will crash the
editor. Learned the hard way.

It's better to just check every time rings are added than waste checks
on every tick I think. However, I think it it would probably be best to
overhaul how rings are added to SonicBP in general. The current
implementation makes other actors get rings directly from SonicBP and
iterate with their BP logic.  In order to be able to run the add lives
at each threshold, every time you want something to add more rings to
the counter you'll need to add the event check right after. This seems
like it can be easily forgotten and thus made a problem for bug checking
anything that manipulates the ring count. What would be better code (and
programming practice) is if SonicBP had a function that handled adding
rings to the counter and then called a threshold check function to see
if any lives were gained. All any other actors would need to do is call
that function on SonicBP and pass how many rings it wanted to add or
subtract.