Ring pick up logic update.
-Changed SonicBP to have an event to handle picking up rings and handle
all the other logic associated with it like adding a 1up at 100 ring
intervals. Makes it easier for other actors to call and pass the amount
of rings to add and not worry about remembering to check for lives.

-Changed Gold_Ring to reflect this, and also tied the number of rings
added on pickup Gold_Ring to a var with a default value of 1 so any new
children of the class could easily change that number like the 5 Ring
and 15 Ring items.

-Changed the Debug Ring to now make use of the changes in Gold_Ring
without having to rewrite the overlap logic. Nice and clean now.

-added more comments in general and tried to make them look nicer