More events/functions, ring collision, sound fixes
-Added more events from previous logic in SonicBP to better reuse and
tidy up the existing functionality. Death is now its own event that
passes a string which signifies its cause, so it can be reused for
various types of death and do different things for those deaths. For
example, a character should not play the knocked back animation if they
die from crushing or death pit. It's still a little messy and I'm pretty
sure passing a string is not the best option here, but it works for now.

-Persistent_Instance_Data now has functions that can be called to modify
the life count. I'm pretty sure this isn't needed given how simple it is
to set to just cast and get the variable but oh well. changeLives takes
in an integer to add to the number of lives. addLife and loseLife use
this function to add/subtract one life. I wanted to leave the option of
adding or subtracting more than one to the amount of lives to more than
one without casting and modifying the variable outside
Persistent_Instance_Data. SonicBP just needs to call addLife or loseLife
now at the relevant events.

-Changed around some var names in the SonicBP and persistentinstancedata
to be more clear

-Changed the ring sound effect to be activated from SonicBP's gain rings
event and removed it from the Gold_Ring actor. This makes it so the
collision overlap doesn't set off the sounds while the rings are
uncollectable and being flung in the Dropped_Ring child. Only a
confirmed pickup sets off the sound.

-Tweaked the jump sound effect so that it plays only when the SonicBP
detects it's touching the ground. This way you can't mash the button and
get the jump sound every time.

-Changed Rings to now use a larger sphere overlap instead of the mesh to
help with collection. Note that scaling the mesh without scaling the
changing collision sphere to be smaller makes it so any children actors
that are larger like the test Super Ring can be collected from quite a
distance away from the mesh. We'll need to tweak the size of rings and
the collision sphere until it feels right.