Combine Ring Powerup and Sound fixes
SonicBP
-Added preliminary BP logic via a function named dropCombineRing in
SonicBP as a simple power up with a boolean var to check the state. When
active the combine ring acts like it does in Knuckles Chaotix and drops
one larger ring that is worth the value of whatever rings were held
before being hurt. The spilled ring nodes were changed to check for the
powerupCombineRing to be active. There is no way to set it in the world
just yet so the only way to test the functionality is to enable the
powerupCombineRing? as true by default. After the initial hit the
function changes the value to false so you only get one go before it
wears off.  Currently the Ring scales in 3 increments based on ring
count and by default is 2 times larger.
-added a few placeholder sounds to the SFX vars so they would play after
the sound file purge.

Combine_Ring
- a child of Dropped_Ring. casts to sonicBP to get the current ring
count when spawned in the world. Would like to see if it is possible to
simply pass the ring_count in on spawn instead of the quick and nasty
casting to SonicBP but I can't seem to find a solution.  The ring by
default is 2 times larger and takes 4 seconds longer to disappear.
Additionally for debug purposes I made a separate material for debugging
before I messed with scaling. Collision radius is naturally smaller so
that you more or less touch the mesh before picking it up.

-Gold_Ring
I messed with it a little should be the same

DEBUG_Super_Ring
-Shrunk sphere collision mesh