SonicBP functionality fixes
SonicBP
-Disabled invincibility on boost and super form additions until we have
a better way to implement that functionality. Boost was changing
invulnerability states every tick resulting in the damage grace
invulnerability being overwritten. It needs a rewrite to an event.
-rewrote the jump logic and readded playerHasJumpControl? since
Ricardo's used the original for double jumps. You should not be able to
jump out of damage or death states until control returns, and double
jump logic was streamlined to be cleaner. Rewritten so that falling off
an object still retains just 1 extra jump and not two.